mysql: drop support for horizontal partitioning of trans/obj
It has never been enabled and the code to drop partitions will be changed in
a way that only 'trans' may still benefit of partitioning. We'll see in the
future if we have cases where 'trans' is too big to delete all rows (of a
given partition) in a single query.